Add: parsing sphere plane  and clean some file
